Moved asset loading functions into assets.go
